requirements.txt
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
aws-sam-translator==1.32.0
aws-xray-sdk==2.6.0
Babel==2.9.0
beautifulsoup4==4.9.1
billiard==3.3.0.23
bleach==3.2.1
boto==2.49.0
boto3==1.16.34
botocore==1.19.34
braintree==4.0.0
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
celery==3.1.26.post2
certifi==2020.6.20
cffi==1.14.4
cfn-lint==0.43.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
cryptography==3.3.1
cssselect2==0.4.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.7
django-admin-thumbnails==0.2.5
django-bootstrap-modal-forms==1.4.1
django-celery==3.3.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-currencies==0.10.1
django-filter==2.4.0
django-import-export==2.3.0
django-inline-actions==2.3.0
django-js-asset==1.2.2
django-leaflet==0.27.1
django-mptt==0.11.0
# Editable install with no version control (django-nested-admin==3.3.2)
-e c:\users\hp\appdata\local\programs\python\python38-32\lib\site-packages
django-nested-inline==0.4.2
django-pyodbc-azure==2.1.0.0
django-pyodbc-gis==0.0.7
django-rest-knox==4.1.0
django-widget-tweaks==1.4.2
django-windows-tools==0.2.1
djangorestframework==3.11.0
docker==4.4.0
docutils==0.16
ecdsa==0.14.1
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
fs==2.4.11
future==0.18.2
geojson==2.5.0
gunicorn==20.0.4
html5lib==1.1
idna==2.9
imagesize==1.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
jsondiff==1.2.0
jsonfield==2.1.1
jsonpatch==1.28
jsonpickle==1.4.2
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
keyring==21.5.0
kombu==3.0.37
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.3
more-itertools==8.6.0
moto==1.3.16
networkx==2.5
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.7
pandas==1.1.1
Pillow==7.1.2
pkginfo==1.6.1
postgis==1.0.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
pylint==2.5.3
pyodbc==4.0.30
pyparsing==2.4.7
Pyphen==0.10.0
pyrsistent==0.17.3
pyshp==2.1.2
python-coveralls==2.9.3
python-dateutil==2.8.1
python-gettext==4.0
python-jose==3.2.0
python-monkey-business==1.0.0
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.12.1
rfc3986==1.4.0
rsa==4.6
s3transfer==0.3.3
schedule==0.6.0
selenium==3.14.0
Shapely==1.7.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
sshpubkeys==3.1.0
tablib==2.0.0
tinycss2==1.1.0
toml==0.10.1
tqdm==4.54.1
twine==3.2.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.29
virtualenvwrapper-win==1.2.6
WeasyPrint==52.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.4.0
